Fix mock managers to have correct return types for collecting PR times
Changed the mock functions to fit the changes made to collect accurate PR configuration times.

+++ tests/dbmstodspi/query_execution/table_data/mock_memory_manager.hpp
@@ -27,7 +27,7 @@ using orkhestrafs::dbmstodspi::MemoryManagerInterface;
 
 class MockMemoryManager : public MemoryManagerInterface {
  public:
-  MOCK_METHOD(int, GetTime, (), (override));
+  MOCK_METHOD(long, GetTime, (), (override));
   MOCK_METHOD(MemoryBlockInterface*, GetAvailableMemoryBlock, (), (override));
   MOCK_METHOD(void, FreeMemoryBlock,
               (MemoryBlockInterface * memory_block_pointer), ());
